Refactor some files into new files
