+++ lib/src/components/MSAView.tsx
@@ -1,4 +1,4 @@
-import React, { useRef, useState, useEffect } from 'react'
+import React from 'react'
 
 import { MsaViewModel } from '../model'
 import { observer } from 'mobx-react'
@@ -10,140 +10,19 @@ import MSACanvas from './MSACanvas'
 import Ruler from './Ruler'
 import TreeRuler from './TreeRuler'
 import Header from './Header'
+import Track from './Track'
 
-const resizeHandleWidth = 5
-
-const Track = observer(
-  ({ model, track }: { model: MsaViewModel; track: any }) => {
-    const { rowHeight, treeAreaWidth } = model
-    return (
-      <div key={track.id} style={{ display: 'flex', height: rowHeight }}>
-        <div style={{ overflow: 'hidden', width: treeAreaWidth }}>
-          <TrackLabel model={model} name={track.name} />
-        </div>
-        <div style={{ width: resizeHandleWidth }} />
-        <track.ReactComponent model={model} {...track} />
-      </div>
-    )
-  },
-)
-
-const TrackLabel = observer(
-  ({ name, model }: { model: MsaViewModel; name: string }) => {
-    const ref = useRef<HTMLCanvasElement>(null)
-    const { rowHeight, treeAreaWidth, highResScaleFactor } = model
-    const width = treeAreaWidth
-
-    useEffect(() => {
-      const canvas = ref.current
-      if (!canvas) return
-      const ctx = canvas.getContext('2d')
-      if (!ctx) return
-
-      ctx.resetTransform()
-      ctx.scale(highResScaleFactor, highResScaleFactor)
-      ctx.clearRect(0, 0, treeAreaWidth, rowHeight)
-      ctx.textAlign = 'right'
-      ctx.font = ctx.font.replace(/\d+px/, `${Math.max(8, rowHeight - 8)}px`)
-      ctx.fillStyle = 'black'
-      ctx.fillText(name, width - 20, rowHeight - rowHeight / 4)
-    }, [name, width, treeAreaWidth, rowHeight])
-    return (
-      <canvas
-        ref={ref}
-        width={width * highResScaleFactor}
-        height={rowHeight * highResScaleFactor}
-        style={{
-          width,
-          height: rowHeight,
-          position: 'relative',
-          overflow: 'hidden',
-        }}
-      />
-    )
-  },
-)
-
-const VerticalResizeHandle = observer(({ model }: { model: MsaViewModel }) => {
-  const [cropMouseDown, setCropMouseDown] = useState(false)
-
-  // this has the effect of just "cropping" the tree area
-  useEffect(() => {
-    if (cropMouseDown) {
-      const listener = (event: MouseEvent) => {
-        model.setTreeAreaWidth(model.treeAreaWidth + event.movementX)
-      }
-
-      const listener2 = () => setCropMouseDown(false)
-
-      document.addEventListener('mousemove', listener)
-      document.addEventListener('mouseup', listener2)
-      return () => {
-        document.removeEventListener('mousemove', listener)
-        document.removeEventListener('mouseup', listener2)
-      }
-    }
-    return () => {}
-  }, [cropMouseDown, model])
-
-  return (
-    <div>
-      <div
-        onMouseDown={() => setCropMouseDown(true)}
-        style={{
-          cursor: 'ew-resize',
-          height: '100%',
-          width: resizeHandleWidth,
-          background: `rgba(200,200,200)`,
-          position: 'relative',
-        }}
-      />
-    </div>
-  )
-})
-
-const HorizontalResizeHandle = observer(
-  ({ model }: { model: MsaViewModel }) => {
-    const [cropMouseDown, setCropMouseDown] = useState(false)
-
-    // this has the effect of just "cropping" the tree area
-    useEffect(() => {
-      if (cropMouseDown) {
-        const listener = (event: MouseEvent) => {
-          model.setHeight(model.height + event.movementY)
-        }
-
-        const listener2 = () => setCropMouseDown(false)
-
-        document.addEventListener('mousemove', listener)
-        document.addEventListener('mouseup', listener2)
-        return () => {
-          document.removeEventListener('mousemove', listener)
-          document.removeEventListener('mouseup', listener2)
-        }
-      }
-      return () => {}
-    }, [cropMouseDown, model])
-
-    return (
-      <div>
-        <div
-          onMouseDown={() => setCropMouseDown(true)}
-          style={{
-            cursor: 'ns-resize',
-            width: '100%',
-            height: resizeHandleWidth,
-            background: `rgba(200,200,200)`,
-            position: 'relative',
-          }}
-        />
-      </div>
-    )
-  },
-)
+import { HorizontalResizeHandle, VerticalResizeHandle } from './ResizeHandles'
 
 export default observer(({ model }: { model: MsaViewModel }) => {
-  const { done, initialized, treeAreaWidth, height, tracks } = model
+  const {
+    done,
+    initialized,
+    resizeHandleWidth,
+    treeAreaWidth,
+    height,
+    tracks,
+  } = model
 
   return (
     <div>
